+++ src/components/Budget/BudgetForm.js
@@ -114,10 +114,10 @@ class BudgetForm extends Component {
                 <Field label="Code" name="code" id="code" component={TextField} fullWidth />
               </Col>
               <Col xs={6}>
-                <Field label="Limit Encumbrance Percent" name="limit_enc_percent" id="limit_enc_percent" component={TextField} type="number" fullWidth />
+                <Field label="Allowable Encumbrance Percent" name="limit_enc_percent" id="allowable_enc_percent" component={TextField} type="number" fullWidth />
               </Col>
               <Col xs={6}>
-                <Field label="Limit Expenditure Percent" name="limit_exp_percent" id="limit_exp_percent" component={TextField} type="number" fullWidth />
+                <Field label="Allowable Expenditure Percent" name="limit_exp_percent" id="allowable_exp_percent" component={TextField} type="number" fullWidth />
               </Col>
               <Col xs={6}>
                 <Field label="Allocation*" name="allocation" id="allocation" component={TextField} validate={[Required]} type="number" fullWidth />