+++ src/components/FiscalYear/FiscalYear.js
@@ -33,7 +33,7 @@ class FiscalYear extends Component {
       initialValue: {
         query: '',
         filters: '',
-        sort: 'Name'
+        sort: ''
       },
     },
     resultCount: { initialValue: INITIAL_RESULT_COUNT },